Implement lag ranges for criticality
Previously, this script just checked if the master and monitored SOAs
were identical, which in a sane world is probably is probably
reasonable. But with dynamic DNS registration in a busy ActiveDirectory
domain (and sometimes flaky Microsoft DNS servers), zones are likely to
lag a bit but not really indicate an abnormal state.

This commit adds -w and -c options which are maximum differences between
master and monitored zone SOAs.